Move test helper functions to `tests/__init__.py`
